🌀 refactor Moved `useInterval.ts` to the hooks directory
